requirements.txt
beautifulsoup4==4.10.0
Django==3.2.9
djangorestframework==3.12.4
requests==2.25.1
selenium==4.0.0
undetected_chromedriver==3.0.3
